[CWS] refactor eventstream
